bc the photo gallery is complete

+++ OurTrips/app/src/main/java/com/alsaeedcullivan/ourtrips/ViewPictureActivity.java
@@ -8,22 +8,33 @@ import android.os.Bundle;
 import android.util.Log;
 import android.view.Menu;
 import android.view.MenuItem;
+import android.view.View;
 import android.widget.ImageView;
+import android.widget.ProgressBar;
+import android.widget.TextView;
 
 import com.alsaeedcullivan.ourtrips.cloud.AccessBucket;
 import com.alsaeedcullivan.ourtrips.cloud.AccessDB;
 import com.alsaeedcullivan.ourtrips.glide.GlideApp;
 import com.alsaeedcullivan.ourtrips.models.Pic;
 import com.alsaeedcullivan.ourtrips.utils.Const;
+import com.google.android.gms.tasks.OnCompleteListener;
 import com.google.android.gms.tasks.Task;
+import com.google.android.gms.tasks.Tasks;
 import com.google.firebase.storage.FirebaseStorage;
 import com.google.firebase.storage.StorageReference;
+
+import java.util.ArrayList;
 
 public class ViewPictureActivity extends AppCompatActivity {
 
     private Pic mPhoto;
     private String mTripId;
-    ImageView mImage;
+    private ImageView mImage;
+    private TextView mLoading;
+    private ProgressBar mSpinner;
+    private int mPosition = -1;
+    private ArrayList<Pic> mPics;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -35,13 +46,22 @@ public class ViewPictureActivity extends AppCompatActivity {
 
         Intent intent = getIntent();
         if (intent != null && intent.getStringExtra(Const.TRIP_ID_TAG) != null &&
-                intent.getParcelableExtra(Const.PIC_TAG) != null) {
+                intent.getParcelableExtra(Const.PIC_TAG) != null &&
+                intent.getIntExtra(Const.POSITION_TAG, -1) != -1
+                && intent.getParcelableArrayListExtra(Const.GALLERY_TAG) != null) {
             mPhoto = intent.getParcelableExtra(Const.PIC_TAG);
             mTripId = intent.getStringExtra(Const.TRIP_ID_TAG);
+            mPosition = intent.getIntExtra(Const.POSITION_TAG, -1);
+            mPics = intent.getParcelableArrayListExtra(Const.GALLERY_TAG);
         } else finish();
 
-        // get a reference to the image view
+        // get a reference to the widgets and set initial visibility
         mImage = findViewById(R.id.view_image_picture);
+        mLoading = findViewById(R.id.view_loading);
+        mSpinner = findViewById(R.id.view_spinner);
+        mLoading.setVisibility(View.GONE);
+        mSpinner.setVisibility(View.GONE);
+        mImage.setVisibility(View.VISIBLE);
 
         // load the picture
         StorageReference ref = FirebaseStorage.getInstance().getReference(mPhoto.getPicPath());
@@ -58,6 +78,10 @@ public class ViewPictureActivity extends AppCompatActivity {
     public boolean onOptionsItemSelected(@NonNull MenuItem item) {
         switch (item.getItemId()) {
             case android.R.id.home:
+                Intent intent = new Intent(ViewPictureActivity.this, GalleryActivity.class);
+                intent.putExtra(Const.GALLERY_TAG, mPics);
+                intent.putExtra(Const.TRIP_ID_TAG, mTripId);
+                startActivity(intent);
                 finish();
                 return true;
             case R.id.delete_photo:
@@ -72,11 +96,37 @@ public class ViewPictureActivity extends AppCompatActivity {
      * deletes this photo from storage and the db
      */
     private void deletePhoto() {
-        if (mPhoto == null || mTripId == null || mPhoto.getDocId() == null) return;
+        if (mPhoto == null || mTripId == null || mPhoto.getDocId() == null || mPosition == -1 ||
+                mPics == null) return;
         Log.d(Const.TAG, "deletePhoto: " + mPhoto.getDocId());
         Log.d(Const.TAG, "deletePhoto: " + mPhoto.getPicPath());
+        // show the progress bar
+        showSpinner();
+        // remove this photo from the list
+        mPics.remove(mPosition);
         // remove the photo from the db and storage
         Task<Void> dbTask = AccessDB.deleteTripPhoto(mTripId, mPhoto.getDocId());
         Task<Void> storeTask = AccessBucket.deleteFromStorage(mPhoto.getPicPath());
+        // when the photo is deleted, finish the activity
+        Tasks.whenAll(dbTask, storeTask).addOnCompleteListener(new OnCompleteListener<Void>() {
+            @Override
+            public void onComplete(@NonNull Task<Void> task) {
+                Intent intent = new Intent(ViewPictureActivity.this, GalleryActivity.class);
+                intent.putExtra(Const.GALLERY_TAG, mPics);
+                intent.putExtra(Const.TRIP_ID_TAG, mTripId);
+                startActivity(intent);
+                finish();
+            }
+        });
+    }
+
+    /**
+     * showSpinner()
+     * shows the progress bar
+     */
+    private void showSpinner() {
+        mImage.setVisibility(View.GONE);
+        mSpinner.setVisibility(View.VISIBLE);
+        mLoading.setVisibility(View.VISIBLE);
     }
 }
